Parser: Fix MSVC warning
Conversion from size_t to uint32_t.

+++ Parsers/CppParser/cppdocumentparser.cpp
@@ -359,8 +359,8 @@ void CppDocumentParser::queueFilesForUpdate()
   static CppTools::CppModelManager *modelManager = CppTools::CppModelManager::instance();
 
   QStringSet filesToUpdate;
-  uint32_t filesOutstanding;
-  uint32_t filesInProcess;
+  size_t filesOutstanding;
+  size_t filesInProcess;
 
   {
     QMutexLocker locker(&d->fileQeueMutex);
